Be consistent with kind/spec in plugin-related definitions
