+++ ui/core/src/model/graph-queries.ts
@@ -11,22 +11,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/**
- * Json config object.
- */
-export interface JsonObject {
-  [key: string]: Json | undefined;
+import { Definition } from './definitions';
+
+export interface GraphQueryDefinition<PluginSpec = unknown> extends Definition<GraphQuerySpec<PluginSpec>> {
+  kind: 'GraphQuery';
 }
 
-/**
- * Json definition values.
- */
-export type Json = string | number | boolean | null | JsonObject | Json[];
-
-/**
- * Base type for definitions in JSON config resources.
- */
-export interface Definition<Spec> {
-  kind: string;
-  spec: Spec;
+export interface GraphQuerySpec<PluginSpec> {
+  graphQueryPlugin: Definition<PluginSpec>;
 }
