+++ ui/core/src/model/panels.ts
@@ -1,4 +1,4 @@
-// Copyright 2021 The Perses Authors
+// Copyright 2022 The Perses Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -11,16 +11,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { Definition, JsonObject } from './definitions';
+import { Definition } from './definitions';
 
-/**
- * Panel definition options that are common to all panels.
- */
-export interface PanelDefinition<Options extends JsonObject = JsonObject> extends Definition<Options> {
+export interface PanelDefinition<PluginSpec = unknown> extends Definition<PanelSpec<PluginSpec>> {
+  kind: 'Panel';
+}
+
+export interface PanelSpec<PluginSpec> {
   display: {
     name: string;
     description?: string;
   };
+  panelPlugin: Definition<PluginSpec>;
 }
 
 /**
