+++ ui/core/src/model/variables.ts
@@ -1,4 +1,4 @@
-// Copyright 2021 The Perses Authors
+// Copyright 2022 The Perses Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -11,56 +11,38 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { Definition, JsonObject } from './definitions';
-
-export const DEFAULT_ALL_VALUE = '$__all' as const;
+import { Definition } from './definitions';
 
 export type VariableName = string;
 
-export type VariableState = {
-  value: VariableValue;
-  options?: VariableOption[];
-  loading: boolean;
-  error?: Error;
-};
-
-export type VariableOption = { label: string; value: string };
-
-export type VariableStateMap = Record<VariableName, VariableState>;
-
 export type VariableValue = string | string[] | null;
 
-export interface Variable<Kind extends string, Options extends JsonObject = JsonObject> extends Definition<Options> {
-  kind: Kind;
+interface VariableSpec {
   name: VariableName;
   display?: {
     label?: string;
     hidden?: boolean;
   };
   defaultValue?: VariableValue;
-  options: Options;
 }
 
-export interface TextVariableOptions extends JsonObject {
+export interface TextVariableDefinition extends Definition<TextVariableSpec> {
+  kind: 'TextVariable';
+}
+
+export interface TextVariableSpec extends VariableSpec {
   value: string;
 }
 
-export type TextVariableDefinition = Variable<'TextVariable', TextVariableOptions>;
+export interface ListVariableDefinition<PluginSpec = unknown> extends Definition<ListVariableSpec<PluginSpec>> {
+  kind: 'ListVariable';
+}
 
-export type ListVariableOptions<Options extends JsonObject> = {
+export interface ListVariableSpec<PluginSpec> extends VariableSpec {
   allowMultiple?: boolean;
   allowAllValue?: boolean;
   customAllValue?: string;
-  optionsLoader: {
-    kind: string;
-    options: Options;
-  };
-};
+  variablePlugin: Definition<PluginSpec>;
+}
 
-export type ListVariableDefinition<Options extends JsonObject = JsonObject> = Variable<
-  'ListVariable',
-  ListVariableOptions<Options>
->;
-
-// All Variables
 export type VariableDefinition = TextVariableDefinition | ListVariableDefinition;
